requirements.txt
sounddevice
scipy
whisper
gtts
python-dotenv
openai